+++ tests/unit/test_api.py
@@ -115,6 +115,7 @@ WEBVIEW_PROPERTIES = (
     "hotkeys_zoom",
     "back_forward_gestures",
     "default_context_menus",
+    "https_scheme",
     "navigation_allow",
     "open_external",
     "download_allow",